facebook/requirements.txt
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
datetime
json5
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
apify-client>=1.4.0 